off_t to size_t change for printf format and malloc. Pointed out by <greg@nest.cx>

+++ sbin/isakmpd/conf.c
@@ -1,4 +1,4 @@
-/*	$OpenBSD: conf.c,v 1.38 2002/04/22 12:52:39 ho Exp $	*/
+/*	$OpenBSD: conf.c,v 1.39 2002/04/29 06:26:50 pvalchev Exp $	*/
 /*	$EOM: conf.c,v 1.48 2000/12/04 02:04:29 angelos Exp $	*/
 
 /*
@@ -574,7 +574,7 @@ conf_reinit (void)
 {
   struct conf_binding *cb = 0;
   int fd, i, trans;
-  off_t sz;
+  size_t sz;
   char *new_conf_addr = 0;
   struct stat sb;
